Create default plan policies on startup, forbid users to modify them

+++ pkg/kore/allocations.go
@@ -19,6 +19,8 @@ package kore
 import (
 	"context"
 
+	"github.com/appvia/kore/pkg/utils/validation"
+
 	configv1 "github.com/appvia/kore/pkg/apis/config/v1"
 	corev1 "github.com/appvia/kore/pkg/apis/core/v1"
 	"github.com/appvia/kore/pkg/store"
@@ -47,7 +49,7 @@ type Allocations interface {
 	// ListAllocationsByType returns a list of all allocations shared to me, filtered by type
 	ListAllocationsByType(ctx context.Context, group, version, kind string) (*configv1.AllocationList, error)
 	// Update is responsible for updating / creating an allocation
-	Update(context.Context, *configv1.Allocation) error
+	Update(ctx context.Context, allocation *configv1.Allocation, allowReadonly bool) error
 }
 
 // acaImpl is the allocations interface
@@ -133,6 +135,10 @@ func (a acaImpl) Delete(ctx context.Context, name string) (*configv1.Allocation,
 		return nil, err
 	}
 
+	if object.Labels[corev1.LabelReadonly] == "true" {
+		return nil, validation.NewError("the allocation can not be deleted").WithFieldError(validation.FieldRoot, validation.ReadOnly, "allocation is read-only")
+	}
+
 	return object, a.Store().Client().Delete(ctx, store.DeleteOptions.From(object))
 }
 
@@ -159,7 +165,7 @@ func (a acaImpl) Get(ctx context.Context, name string) (*configv1.Allocation, er
 func (a acaImpl) GetAssigned(ctx context.Context, name string) (*configv1.Allocation, error) {
 	list, err := a.ListAllocationsAssigned(ctx)
 	if err != nil {
-		log.WithError(err).Error("trying to retrieve list of assigned allocations")
+		log.WithError(err).Error("failed to retrieve list of assigned allocations")
 
 		return nil, err
 	}
@@ -182,7 +188,7 @@ func (a acaImpl) ListAllocationsAssigned(ctx context.Context) (*configv1.Allocat
 		store.ListOptions.InAllNamespaces(),
 		store.ListOptions.InTo(all),
 	); err != nil {
-		log.WithError(err).Error("trying to retrieve a list of all allocations")
+		log.WithError(err).Error("failed to retrieve a list of all allocations")
 
 		return nil, err
 	}
@@ -215,7 +221,7 @@ func (a acaImpl) List(ctx context.Context) (*configv1.AllocationList, error) {
 }
 
 // Update is responsible for updating / creating an allocation
-func (a acaImpl) Update(ctx context.Context, allocation *configv1.Allocation) error {
+func (a acaImpl) Update(ctx context.Context, allocation *configv1.Allocation, allowReadonly bool) error {
 	logger := log.WithFields(log.Fields{
 		"group":              allocation.Spec.Resource.Group,
 		"kind":               allocation.Spec.Resource.Kind,
@@ -240,6 +246,20 @@ func (a acaImpl) Update(ctx context.Context, allocation *configv1.Allocation) er
 	// picked up the controller anyhow
 	if allocation.Spec.Resource.Namespace != a.team {
 		return ErrNotAllowed{message: "you cannot allocate a resource which you do not own"}
+	}
+
+	if !allowReadonly {
+		original, err := a.Get(ctx, allocation.Name)
+		if err != nil && err != ErrNotFound {
+			return err
+		}
+
+		if original != nil && original.Labels[corev1.LabelReadonly] == "true" {
+			return validation.NewError("the allocation can not be updated").WithFieldError(validation.FieldRoot, validation.ReadOnly, "allocation is read-only")
+		}
+		if allocation.Labels[corev1.LabelReadonly] == "true" {
+			return validation.NewError("the allocation can not be updated").WithFieldError(validation.FieldRoot, validation.ReadOnly, "read-only flag can not be set")
+		}
 	}
 
 	return a.Store().Client().Update(ctx,
